Add initial game skeleton
This commit adds the initial skeleton to hook into the Amethyst game
engine. This includes an initial set of states, support for passing
events with custom data, and having type-safe input processing.

As part of this effort there were some issues identified with the
Amethyst dependencies including them using some unexpected licenses and
multiple versions of libraries used. Each of these issues was analyzed
and concluded that these do not prevent us from moving forward. The
cargo-deny configuration and the game design document are updated
to reflect these items.