Add local player and ai player systems
The local player and AI player systems are added along with updating
the components, resources, and updating the Game state to support these
initial systems.

These systems are available via a game bundle.

Clippy was unhappy with the way systems specify the types of data they
wish to receive, thus the type_complexity check is disabled for just
these lines.