Support building the user manual
This commit adds support for building the user manual. The build process
uses mdbook to generate the manual. cargo-about is used to generate the
license information and software BOM sections.

Unfortunately, cargo-about makes it difficult to run in a cross platform
way by requiring it be run from a specific directory and its output be
redirected to a file. Therefore, a helper application is added that runs
cargo-about taking care of doing the output redirection and such. Later
this tool could be extended to handel the asset licenses tool.

Note: currently there is not much actual content in the manual.