Add debug environment
Adding the debug environment includes adding support for the debug lines
system, converting between screen space, world space, and grid positions,
and ensuring the debug camera does not interfere with the ability to
draw menus.

To add the debug environment the grid is updated to support converting
between board positions and world coordinates. Tests are added to ensure
the math is done correctly.

The mouse raycast system is added to track the mouse giving connivent
access to the position in terms of world coordinates and board position.
this keeps the logic in the the player input system small and allows
other systems to track the mouse position, e.g. the hover over grid
cell effect.

Finally, the GameData resource is renamed to GameLogic. States are
already provided a 'GameData' parameter by the engine thus having a
resource with the same name is confusing. Additionally, GameLogic better
describes the purpose of the resource.

The result of this commit is a single game can be fully played.