Add environments support
The basic logic for supporting environments is added. This includes the
environment trait, the environments collection, and an initial debug
environment to try the logic.

Because environments need access to the world, the world cannot contain
the environments during modifications --- otherwise borrow errors occur.
Therefore, the environments resource is an optional: users can pull the
environments out of the world, modify as needed, then replace when done.