Add basic main menu
The main menu contains the buttons necessary for the first milestone and
support for handling button events.

The main menu is defined in a .ron file at this time as this was the
simplest way to move forward. However, the .ron format does not support
shared styles so this might become a maintenance issue in the future.
To address we could consider how to construct the UI in code, or figure
out how to apply a style to a tree of UI elements.

Currently, ony the transition back to the game state and application
exit are supported.