Attempt to get menus to work
This commit updates the single player game so it shows status text and
the menu button.

I encountered a bug where opening the menu would cause some menu buttons
to be placed in weird spots and persist ... an alternate approach might
need to be taken for handling the UI.

This commit also has additional cleanup such as fixing warnings and
adding some delay to the AI player.